+++ audiostack/__init__.py
@@ -4,7 +4,6 @@ api_key = None
 Authorization = None
 assume_org_id = None
 app_info = None
-customer_trace_id = None
 
 TIMEOUT_THRESHOLD_S = 300